Pr 301 tweaks (#312)
* move (very nice) router example from Servers.jl to Routers.jl; move handle(h::RequestHandler, stream::Stream) to Servers.jl and misc reordering / whitespace changes to reduce diff site

* Revert stream::HTTP.Stream local variable name to http::HTTP.Stream.
This is consistent with with Streams.jl and StreamRequest.jl.
Also removes inconsistency with stream::Bool in `listen`.

* Make HandlerFunction type callable.
e.g. (h::RequestHandlerFunction)(req) = handle(h,req)

This allows `listen` to treat the `h` handler parameter is a compeltely
generic way. i.e. functions, or Handlers or anything else that uses
the `object(args)` call pattern can be used as a request handler.

When `stream::Bool=true` the `::Stream` is passed to the handler `h` directly.
When `stream::Bool=false` the handler is wrapped a call to `handle(h, ::Stream)`.
