+++ src/Routers.jl
@@ -1,3 +1,207 @@
+"""
+# Examples
+Let's put together an example http REST server for our hypothetical "ZooApplication" that utilizes various
+parts of the Servers & Handlers frameworks.
+
+Our application allows users to interact with custom "animal" JSON objects.
+
+First we have our "model" or data structures:
+```julia
+mutable struct Animal
+    id::Int
+    type::String
+    name::String
+end
+```
+
+Now we want to define our REST api, or how do we allow users to create, update,
+retrieve and delete animals:
+```julia
+# use a plain `Dict` as a "data store"
+const ANIMALS = Dict{Int, Animal}()
+const NEXT_ID = Ref(0)
+function getNextId()
+    id = NEXT_ID[]
+    NEXT_ID[] += 1
+    return id
+end
+
+# "service" functions to actually do the work
+function createAnimal(req::HTTP.Request)
+    animal = JSON2.read(IOBuffer(HTTP.payload(req)), Animal)
+    animal.id = getNextId()
+    ANIMALS[animal.id] = animal
+    return HTTP.Response(200, JSON2.write(animal))
+end
+
+function getAnimal(req::HTTP.Request)
+    animalId = HTTP.URIs.splitpath(req.target)[5] # /api/zoo/v1/animals/10, get 10
+    animal = ANIMALS[animalId]
+    return HTTP.Response(200, JSON2.write(animal))
+end
+
+function updateAnimal(req::HTTP.Request)
+    animal = JSON2.read(IOBuffer(HTTP.payload(req)), Animal)
+    ANIMALS[animal.id] = animal
+    return HTTP.Response(200, JSON2.write(animal))
+end
+
+function deleteAnimal(req::HTTP.Request)
+    animalId = HTTP.URIs.splitpath(req.target)[5] # /api/zoo/v1/animals/10, get 10
+    delete!(ANIMALS, animal.id)
+    return HTTP.Response(200)
+end
+
+# define REST endpoints to dispatch to "service" functions
+const ANIMAL_ROUTER = HTTP.Router()
+HTTP.@register(ANIMAL_ROUTER, "POST", "/api/zoo/v1/animals", createAnimal)
+HTTP.@register(ANIMAL_ROUTER, "GET", "/api/zoo/v1/animals/*", getAnimal)
+HTTP.@register(ANIMAL_ROUTER, "PUT", "/api/zoo/v1/animals", updateAnimal)
+HTTP.@register(ANIMAL_ROUTER, "DELETE", "/api/zoo/v1/animals/*", deleteAnimal)
+```
+
+Great! At this point, we could spin up our server and let users start managing their animals:
+```julia
+HTTP.listen(ANIMAL_ROUTER, Sockets.localhost, 8081)
+```
+
+Now, you may have noticed that there was a bit of repitition in our "service" functions, particularly
+with regards to the JSON serialization/deserialization. Perhaps we can simplify things by writing
+a custom "JSONHandler" to do some of the repetitive work for us.
+```julia
+function JSONHandler(req::HTTP.Request)
+    # first check if there's any request body
+    body = IOBuffer(HTTP.payload(req))
+    if eof(body)
+        # no request body
+        response_body = handle(ANIMAL_ROUTER, req)
+    else
+        # there's a body, so pass it on to the handler we dispatch to
+        response_body = handle(ANIMAL_ROUTER, req, JSON2.read(body, Animal))
+    end
+    return HTTP.Response(200, JSON2.write(response_body))
+end
+
+# **simplified** "service" functions
+function createAnimal(req::HTTP.Request, animal)
+    animal.id = getNextId()
+    ANIMALS[animal.id] = animal
+    return animal
+end
+
+function getAnimal(req::HTTP.Request)
+    animalId = HTTP.URIs.splitpath(req.target)[5] # /api/zoo/v1/animals/10, get 10
+    return ANIMALS[animalId]
+end
+
+function updateAnimal(req::HTTP.Request, animal)
+    ANIMALS[animal.id] = animal
+    return animal
+end
+
+function deleteAnimal(req::HTTP.Request)
+    animalId = HTTP.URIs.splitpath(req.target)[5] # /api/zoo/v1/animals/10, get 10
+    delete!(ANIMALS, animal.id)
+    return ""
+end
+```
+
+And we modify slightly how we run our server, letting our new `JSONHandler` be the entry point
+instead of our router:
+```julia
+HTTP.listen(JSONHandler, Sockets.localhost, 8081)
+```
+
+Our `JSONHandler` is nice because it saves us a bunch of repitition: if a request body comes in,
+we automatically deserialize it and pass it on to the service function. And each service function
+doesn't need to worry about returning `HTTP.Response`s anymore, but can just focus on returning
+plain Julia objects/strings. The other huge advantage is it provides a clean separation of concerns
+between the "service" layer, which should really concern itself with application logic, and the
+"REST API" layer, which should take care of translating between a web data format (JSON).
+
+Let's take this one step further and allow multiple users to manage users, and add in one more
+custom handler to provide an authentication layer to our application. We can't just let anybody
+be modifying another user's animals!
+```julia
+# modified Animal struct to associate with specific user
+mutable struct Animal
+    id::Int
+    userId::Base.UUID
+    type::String
+    name::String
+end
+
+# modify our data store to allow for multiple users
+const ANIMALS = Dict{Base.UUID, Dict{Int, Animal}}()
+
+# creating a user returns a new UUID key unique to the user
+createUser(req) = Base.UUID(rand(UInt128))
+
+# add an additional endpoint for user creation
+HTTP.@register(ANIMAL_ROUTER, "POST", "/api/zoo/v1/users", createUser)
+# modify service endpoints to have user pass UUID in
+HTTP.@register(ANIMAL_ROUTER, "GET", "/api/zoo/v1/users/*/animals/*", getAnimal)
+HTTP.@register(ANIMAL_ROUTER, "DELETE", "/api/zoo/v1/users/*/animals/*", deleteAnimal)
+
+# modified service functions to account for multiple users
+function createAnimal(req::HTTP.Request, animal)
+    animal.id = getNextId()
+    ANIMALS[animal.userId][animal.id] = animal
+    return animal
+end
+
+function getAnimal(req::HTTP.Request)
+    paths = HTTP.URIs.splitpath(req.target)
+    userId = path[5] # /api/zoo/v1/users/x92jf-.../animals/10, get user UUID
+    animalId = path[7] # /api/zoo/v1/users/x92jf-.../animals/10, get 10
+    return ANIMALS[userId][animalId]
+end
+
+function updateAnimal(req::HTTP.Request, animal)
+    ANIMALS[animal.userId][animal.id] = animal
+    return animal
+end
+
+function deleteAnimal(req::HTTP.Request)
+    paths = HTTP.URIs.splitpath(req.target)
+    userId = path[5] # /api/zoo/v1/users/x92jf-.../animals/10, get user UUID
+    animalId = path[7] # /api/zoo/v1/users/x92jf-.../animals/10, get 10
+    delete!(ANIMALS[userId], animal.id)
+    return ""
+end
+
+# AuthHandler to reject any unknown users
+function AuthHandler(req)
+    if HTTP.hasheader(req, "Animal-UUID")
+        uuid = HTTP.header(req, "Animal-UUID")
+        if haskey(ANIMALS, uuid)
+            return JSONHandler(req)
+        end
+    end
+    return HTTP.Response(401, "unauthorized")
+end
+```
+
+And our mofidified server invocation:
+```julia
+HTTP.listen(AuthHandler, Sockets.localhost, 8081)
+```
+
+Let's review what's going on here:
+  * Each `Animal` object now includes a `UUID` object unique to a user
+  * We added a `/api/zoo/v1/users` endpoint for creating a new user
+  * Each of our service functions now account for individual users
+  * We made a new `AuthHandler` as the very first entry point in our middleware stack,
+    this means that every single request must first pass through this authentication
+    layer before reaching the service layer. Our `AuthHandler` checks that the user
+    provided our security request header `Animal-UUID` and if so, ensures the provided
+    UUID corresponds to a valid user. If not, the `AuthHandler` returns a 401
+    HTTP response, signalling that the request is unauthorized
+
+Voila, hopefully that helps provide a slightly-more-than-trivial example of utilizing the
+HTTP.Handlers framework in conjuction with running an HTTP server.
+"""
+
 module Routers
 
 export Router, @register, register!
